backend/requirements.txt
kafka-python-ng
pymongo
Flask
Flask-Cors
Flask-SocketIO
python-socketio
python-engineio
timesynth
numpy
scikit-learn